[#46] cmake: add support for CMake install
Adds support for `cmake --install` of the entire project including any dependencies. This is a great starting-off point for supporting phoenix to be installed in systems as a shared library which is currently not possible and will be introduced at a later point.

+++ include/phoenix/script.hh
@@ -829,7 +829,7 @@ namespace phoenix {
 			sym._m_type = datatype::string;
 			sym._m_count = 1;
 			sym._m_value = std::unique_ptr<std::string[]> {new std::string[sym._m_count]};
-			sym._m_index = _m_symbols.size();
+			sym._m_index = static_cast<std::uint32_t>(_m_symbols.size());
 
 			return &_m_symbols.emplace_back(std::move(sym));
 		}
